- Merged ViewExt.kt into one file - Minor changes due to PR comments
