+++ app/src/main/java/space/taran/arknavigator/mvp/view/item/FileItemViewHolder.kt
@@ -4,7 +4,6 @@ import android.content.res.ColorStateList
 import android.util.Log
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
-import kotlinx.android.extensions.LayoutContainer
 import space.taran.arknavigator.R
 import space.taran.arknavigator.databinding.ItemFileGridBinding
 import space.taran.arknavigator.ui.fragments.utils.Preview
