--- app/src/main/java/space/taran/arknavigator/utils/ViewExt.kt
@@ -1,16 +1,0 @@
-package space.taran.arknavigator.utils
-
-import android.view.View
-
-fun View.makeGone(){
-    visibility = View.GONE
-}
-
-fun View.makeVisible(){
-    visibility = View.VISIBLE
-}
-
-fun View.makeVisibleAndSetOnClickListener(action: () -> Unit){
-    setOnClickListener{ action() }
-    visibility = View.VISIBLE
-}
